+++ src/App.tsx
@@ -3,7 +3,8 @@ import { Show } from "solid-js";
 import { Editor } from "./Editor";
 import { Region } from "./Region";
 import { Texture } from "./Texture";
-import { createDnd, createResize } from "./hooks";
+import { createDnd } from "./hooks/createDnd";
+import { createResize } from "./hooks/createResize";
 import { AppStoreProvider, useAppStore } from "./store";
 import { EditorToolbar } from "./toolbar/EditorToolbar";
 import { TextureToolbar } from "./toolbar/TextureToolbar";
