+++ src/hooks/createDnd.ts
@@ -27,31 +27,3 @@ export function createDnd(setFile: (blob: Blob) => void) {
 
   return { isDragOver, onDrop, onDragEnter, onDragOver, onDragLeave };
 }
-
-export function createResize() {
-  const [dragging, setDragging] = createSignal(false);
-  const [width, setWidth] = createSignal(50);
-  const left = () => width();
-  const right = () => 100 - width();
-
-  function onMouseMove(e: MouseEvent) {
-    if (dragging()) {
-      const width = (e.clientX / window.innerWidth) * 100;
-      setWidth(width);
-    }
-  }
-
-  function reset() {
-    setWidth(50);
-  }
-
-  function activate() {
-    setDragging(true);
-  }
-
-  function deactivate() {
-    setDragging(false);
-  }
-
-  return { width, left, right, onMouseMove, reset, activate, deactivate };
-}
